Implemented majority voting for "checksum" and "fp_checksum" to protect against modifications that result from fault injections. Majority voting enables us to restore the correct value of "checksum" and "fp_checksum".
TODO: Decide whether majority voting leads to desired behaviour.
Perhaps fault injections that modify the "checksum" or "fp_checksum"
directly should be regarded as "INVALID".

+++ encode/tests/mylibs/mycheck.h
@@ -10,7 +10,7 @@ __uint128_t __cs_acc(__uint128_t x);
 void        __cs_msg(void);
 
 void        __cs_fopen(int argc, char** argv);
-size_t      __cs_facc(uint64_t x);
+size_t      __cs_facc(__uint128_t x);
 void        __cs_fclose();
 
 #endif /* __MYCHECK_H__ */
